test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
hacking!=0.13.0,<0.14,>=0.12.0 # Apache-2.0

fixtures>=3.0.0 # Apache-2.0/BSD
testrepository>=0.0.18 # Apache-2.0/BSD
testscenarios>=0.4 # Apache-2.0/BSD
testtools>=2.2.0 # MIT
oslotest>=3.2.0 # Apache-2.0

# when we can require tox>= 1.4, this can go into tox.ini:
#  [testenv:cover]
#  deps = {[testenv]deps} coverage
coverage!=4.4,>=4.0 # Apache-2.0

# this is required for the sphinx extension
sphinx!=1.6.6,!=1.6.7,>=1.6.2 # BSD

# mocking framework
mock>=2.0.0 # BSD
requests_mock>=1.5.0 # Apache-2.0

# Bandit security code scanner
bandit>=1.1.0 # Apache-2.0
